openflow: Separate flow stats headers from their bodies.
For some time, Open vSwitch has combined both ofp_stats_msg and
the data that follows it into single structures, because we found
that this is usually easier to handle.than having them separated.
But OpenFlow 1.1+ has a different stats message header (it has 4
extra bytes of padding) whereas many of the bodies are the same
or similar and so this approach falls down there.  Therefore, this
commit switches back to separated header and body for stats
messages.

Signed-off-by: Ben Pfaff <blp@nicira.com>