nx-match: Take into account leading header when calculating pad
In the case of Open Flow 1.2, which is currently the only
time that OXM is be used, there is a 4 byte header before
the match which needs to be taken into account when calculating
the pad length.

This is not entirely pretty, but it does seem to be correct.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v5
* Move nx_padded_match_len into nx-match.h as it will
  be used by code outside of nx-match.c added by subsequent patches.

v4
* Pad is needed even if match is zero length
  (what was I thinking?!)

v3
* Initial post