Add Nicira vendor extension action NXAST_DEC_TTL_CNT_IDS.
Currently, if a controller having a nonzero id registers to get a
OFPR_INVALID_TTL async message, it will not receive it.  This is because
compose_dec_ttl() only sent the invalid ttl packets to the default controller
id.  NXAST_DEC_TTL_CNT_IDS is a new action that accepts a list of controller
ids, each separated by `,', to which the OFPR_INVALID_TTL packets must be sent.
The earlier requirement of the controller having to explicitly register to
receive these asynchronous messages is retained.
The syntax of this action is:
    dec_ttl(id1,id2)
where id1, id2 are valid controller ids.

Signed-off-by: Mehak Mahajan <mmahajan@nicira.com>
(cherry picked from commit c2d967a562aaa57b679c302a5041d5a941154143)

Signed-off-by: Simon Horman <horms@verge.net.au>

Conflicts:
	include/openflow/nicira-ext.h
	lib/ofp-actions.h
	lib/ofp-parse.c
	lib/ofp-util.def
	ofproto/ofproto-dpif.c
	tests/ofp-actions.at