datapath: Correctly set key->vlan.type for ETH_P_8021Q
Empirically on linux 3.5, it appears that in the case where
vlan_tx_tag_present() is true skb->protocol has
the ethernet type of the frame after the outer VLAN header is
stripped. This is not the value of vlan.type stored in the key in the flow
table and the result is unnecessary upcalls to ovs-vswitchd.

In the case where vlan_tx_qinq_tag_present() it should be valid to
assume that the type of the VLAN header was ETH_P_8021Q. And in that
case, if an inner tag is present then set the tpid of the skb to
ETH_P_8021Q so that it will be used if the skb hits ovs_flow_extract()
again in which case vlan_tx_qinq_tag_present() will be true.

The other case where vlan_tx_qinq_tag_present() is true is that
the qinq tag has been extracted by other means. In this case
assume that the type of the outer vlan tag is ETH_P_8021AD.
I am entirely unsure about the validity of this assumption,
or what bugs may lie in q-in-q handling code.

Signed-off-by: Simon Horman <horms@verge.net.au>