datapath: Include inner-frame information in MPLS frame flow
This is necessary to provided adequately fine-grained matches for
facets that have actions modify inner-frame data.

e.g. A facet that modifies the IP TTL of an MPLS-encapsulated frame
     should not act on a frame if it is MPLS-encapsulated ARP,

It also provides symmetry with ovs-vswitchd's view of flows.
ovs-vswitchd requires inner-frame information in order to compose
actions that modify inner-frame data. e.g. an action that modifies
the IP TTL, is composed as a SET_IPV4 action and composing this
requires knowledge of the IPv4 addresses of the frame.

Signed-off-by: Simon Horman <horms@verge.net.au>