ofproto: Allow richer matches based on actions
If the actions of a flow allow further decoding of L3 and L4 data to
provide a richer match then use this richer match.

For example:

In the case of MPLS, L3 and L4 information may not initially be decoded
from the frame as the ethernet type of the frame is an MPLS type and no
information is known about the type of the inner frame.

However, the type of the inner frame may be provided by an mpls_pop action
in which case L3 and L4 information may be decoded providing a finer
grained match than is otherwise possible.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v2.19
* As suggested by Ben Pfaff
  - Obtain decode inner_flow in xlate_actions() as it is able to take
    into account resubmit actions.
* Correct typo in comment in flow_miss_is_singleton()
* Update for removal of encap_dl_type parameter from odp_flow_key_from_flow()
* Squashed the non-odp_flow_key_from_flow portions of
  "actions: Allow secondary decoding of a flow"
* Changed patch name from
  "ofproto: Allow actions richer matches based on actions" to
  "ofproto: Allow richer matches based on actions"

v2.16 - v2.18
* No change

v2.15
* Rebase for ovs_assert

v2.14
* No change

v2.13
* Correct indentation in handle_flow_miss_l3_extraction()
* Rebase: Pass encap_dl_type argument to flow_extract_l3_onwards()

v2.12
* Use eth_type_mpls helper
* As suggested by Jarno Rajahalme
  - Use flow->encap_dl_type instead of of obtaining the value
    by scanning the actions.

v2.11
* First post