Track ethertype changes for learn_parse()
learn_parse() needs to know the current ethertype, including any
updates made by MPLS push and pop actions in order to determine if
access to fields is valid or not.

Changes made to ethertype my VLAN push and pop actions are excluded
from this logic as they do not effect the behaviour of learn_parse.

Signed-off-by: Simon Horman <horms@verge.net.au>