ofproto-dpif: Add 'force-miss-model' configuration
This adds support for specifying flow miss handling behaviour at
runtime, through a new "other-config" option in the Open_vSwitch table.
This is an extension to flow-eviction-threshold.

By default, the behaviour is the same as before. If force-miss-model is
set to 1, then flow miss handling will always result in the creation of
new facets and flow-eviction-threshold will be ignored. If
force-miss-model is set to 2, then flow miss handling will never result
in the creation of new facets (effectively the same as setting the
flow-eviction-threshold to 0, which is not currently configurable).

We intend to use this configuration option in the testsuite to force
particular code paths to be used, allowing us to improve test coverage.

Signed-off-by: Joe Stringer <joe@wand.net.nz>
Signed-off-by: Simon Horman <horms@verge.net.au>

---

v13
* First post