Do not update packet->l3 pop_mpls
pop_mpls() modifies packet->data before packet->l3 and thus
packet->l3 should not be modified as it is already correct.

Without this change subsequent l3 changes packet, for example
updating the IP TTL, would result in a corrupted packet as
an incorrect packet-l3 is used as the basis for such changes.

Signed-off-by: Simon Horman <horms@verge.net.au>