ofproto: Allow flow-dump matching on mpls label using OpenFlow 1.2
OpenFlow 1.2 supports MPLS and should allow flow-dump requests
to match on the on the MPLS label.

Without this change a request to do so will be downgraded to
NXT, which will fail because the osrg tree only permits
OpenFlow 1.2 connections.

Also allow matching on the mpls stack bit using OpenFlow 1.2

Signed-off-by: Simon Horman <horms@verge.net.au>