ofp-actions: Correct handling of apply actions instruction
When stored in OFP actions the apply actions instruction is represented using
nested attributes.  This means that using OFPACT_FOR_EACH() will skip
to the next instruction as desired.

Previously the code assumed that nesting was not used and that
OFPACT_FOR_EACH() would merely skip to the second action inside the apply
actions instruction. To compensate for this ofpacts_insts_to_openflow11__()
returned the last OFP action that had been processed - the last action in
the case of apply actions instruction - and then skip to the next action.
Unfortunately this was merely the end marker of the nested attribute and
processing would end: prematurely if there were more instructions.

This patch simplifies the code by simply using OFPACT_FOR_EACH()
to skip over nested attributes.

This patch also adds a test case to exercise this condition.

Reported-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>
Signed-off-by: Simon Horman <horms@verge.net.au>