ofp-actions: Stop processing of entire table on invalid TTL decrement
In the case of decrementing to an invalid NW or MPLS TTL processing of
the table should be terminated. However do to the nested processing of
the apply actions table only the processing of the apply actions
instruction was terminated and subsequent instructions for the table
would be processed.

For instance in the following case the packet would be processed
to table 2 even if dec_ttl was invalid.

ovs-ofctl add-flow br0 "icmp actions=dec_ttl,goto_table:2"

This patch resolves this problem by adding an exit_table
field to struct action_xlate_ctx and using it to determine if
table processing should continue or not.

This problem is specific to the osrg tree.

Signed-off-by: Simon Horman <horms@verge.net.au>