ofp-actions: Allow resubmit and goto_table after MPLS push
Background:

In order to supply some visibility to the inner-packet of
an MPLS packet the encap_dl_type field was added to struct flow
and is set in several places including when translating an
push MPLS action.

This turns out to be a poor design because the inner type can only
be known at translation time and a packet may not be available at this point:
in particular a packet is not available when translation occurs for
flow validation. Yet without a packet its not possible to decode fields
of the inner protocol. This approach has been abandoned upstream in favour
of recirculation but remains in the OSRG tree.

Problem:

One side-effect of this field being present is that if it
is a non-zero value then no rule will match, as no rules have
this field set. So resubmit and goto_table do not match any rules
that include any match at all: that is any rule that does not
use all-wildcards.

Fix:

In order to work around this the field is zeroed and restored
when looking up rules for resubmit and goto_table.

Signed-off-by: Simon Horman <horms@verge.net.au>