Remove mpls_depth field from flow
Rather than tracking the MPLS depth as a field in the
flow, which is an entirely poor place for it, just track
the delta to the MPLS depth during translation.

This logic was developed while implementing recirculation
and intended to be used to detect when recirculation should
occur. This variant of the patch uses the logic to determine
if processing of actions should stop due to an MPLS
action which cannot be translated (without recirculation).

A side-effect of this patch is that it resolves a bug
whereby ovs-vswitchd will abort due to to an assertion
on eth_type_mpls(ctx->xin->flow.dl_type) in compose_mpls_pop_action(()
if the actions of a flow include pop_mpls twice without
a push_mpls in between.

Signed-off-by: Simon Horman <horms@verge.net.au>

---
v3
* As suggested by Ben Pfaff
  - Remove empty zeros field from struct flow

v2
* As suggested by Ben Pfaff
  - Move pre_push_mpls_lse fron a local variable of do_xlate_actions()
    to a field of struct xlate_ctx so that it is accessible across resubmit
    and goto table actions.
  - Move removal of eth_mpls_depth() into a separate patch.
    It is not strictly related to this change.
  - Correct spelling errors