ofproto: Honour Table Mod settings for table-miss handling
This reworks lookup of rules for both table 0 and
table action translation. The result is that Table Mod settings,
which can alter the miss-behaviour of tables, including table 0,
on a per-table basis may be honoured.

Previous patches proposed by myself which build
on earlier merged patches by Andy Zhou implement the ofproto
side of Table Mod. So with this patch the feature should be complete.

This patch, nor any other patches it builds on, alter the default
behaviour of Open vSwitch. And in particular the OpenFlow1.1 behaviour
is the default regardless of which OpenFlow version is negotiated
between the switch and the controller.

An implementation detail, which lends itself to future work, is
the handling of OFPTC_TABLE_MISS_CONTINUE. If a table has this
behaviour set by Table Mod and a miss occurs then a loop is created,
skipping to the next table. It is quite easy to create
a situation where this loop covers ~255 tables which is very
expensive as the lookup for each table involves taking locks,
amongst other things. One way to do this would be to maintain
a bitmap, one bit per table, to indicate if the table has
any rules present or not.

Cc: Andy Zhou <azhou@nicira.com>
Signed-off-by: Simon Horman <horms@verge.net.au>