netdev-linux: Read packet auxdata to obtain vlan_tid
If VLAN acceleration is used when the kernel receives a packet
then the outer-most VLAN tag will not be present in the packet
when it is received by netdev-linux. Rather, it will be present
in auxdata.

This patch uses recvmsg() instead of recv() to read auxdata for
each packet and if the vlan_tid is set then it is added to the packet.

Adding the vlan_tid makes use of headroom available
in the buffer parameter of rx_recv.

Co-authored-by: Ben Pfaff <blp@nicira.com>
Signed-off-by: Simon Horman <horms@verge.net.au>

--
v5
* As provided by Ben Pfaff
  - Include local copy of struct tpacket_auxdata and related #defines
    to avoid depending on the kernel headers used for the build.
* Include update of documentation of rx_recv which
  was previously included in an earlier patch of the series.

v4
* Rebase
* Make aux variable in netdev_linux_rx_recv_sock() const
* Use TP_STATUS_VLAN_VALID to allow vlan 0 to be read from auxdata
* Use TP_STATUS_VLAN_TPID_VALID allow the VLAN tpid to be read
  from auxdata

v3
* Remove unnecessary cast
* On success return 0 and increment buffer->size
* Limit to reading one VLAN from auxdata

v2
* Updated to use eth_push_vlan() to efficiently push the VLAN
  if headroom is available. This is possible now there
  is a struct ofpbuf * to work with.
* Various cleanups