ofp-actions: Expose position argument of push_mpls to ovs-ofctl
In the presence of VLAN tags there are two possibilities for pushing
an MPLS LSE described in different versions of OpenFlow: before or after
the VLAN tags.

* Pushing after any VLAN tags that are present is the behaviour described in
  OpenFlow1.1 and 1.2 and the behaviour that Open vSwitch currently uses
  for all OpenFlow versions.

* Pushing before any VLAN tags that are present is the behaviour described
  in OpenFlow1.3

The difference the between these can alter the consistency of actions.

For example, if an MPLS LSE is pushed before VLAN tags then it is valid to
subsequently push a VLAN tag. This is because the body of an MPLS packet is
opaque and thus after an MPLS LSE a VLAN packet becomes a non-VLAN packet
for the purposes of action consistency checking.

Conversely, if an MPLS LSE is pushed after VLAN tags then it is not valid
to subsequently push a VLAN tag because Open vSwitch does not support
pushing a VLAN tag onto a VLAN packet.

As suggested by Ben Pfaff, in order to allow disambiguation of these cases
a position parameter to push_mpls is exposed to ovs-ofctl by this patch.

Currently the default and only supported value of position is
'before_vlan'. This requires OpenFlow1.2 or earlier. And an implication of
this patch is that push_mpls is prohibited for OpenFlow1.3 and later.

Prior to this patch push_mpls was allowed for OpenFlow1.3 and later
but its implementation did not follow the specification in the presence
of VLANS: the MPLS LSE was pushed after rather than before them.

A subsequent patch enables pushing the MPLS before VLANs and
adds allows 'after_vlan' as a value of the position parameter
to push_mpls. 'after_vlan' will require OpenFlow1.3 or later.

This patch has been broken out in order to aid review.
It can trivially be combined with the subsequent patch if desired.

This patch also:
* Updates consistency checking to be dependent on the position
  argument of push_mpls actions.
* Updates the test suite to use of the position argument
  for mpls_push actions.

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>

---
v2.52
* First post

  This approach was suggested by Ben Pfaff.

  He also suggested adding a 'no_vlan' position which would be valid for
  all OpenFlow versions if no VLAN tag was present.

  I have not included that in this patch because I am not clear on how to
  handle it in conjunction with dumping flows. It seems to me that if
  OpenFlow 1.2 or earlier it would be translated to 'after_vlan' while if
  (in the presence of the subsequent patch described above) OpenFlow 1.3 or
  later was used it would be translated to 'before_vlan'.

  This is because ultimately flows are dumped using OpenFlow which does not
  have a position parameter. So it would be set when reading OpenFlow
  messages  according to whichever OpenFlow version was used to dump the
  flows.

  It seems to me that this would be rather confusing for users.

  Other than the problem described immediately above implementing 'no_vlan'
  seems trivial and I had it working at one stage of my development of this
  patch.