netdev_class: Pass a struct ofpbuf * to rx_recv()
Update the netdev_class so that struct ofpbuf * is passed to rx_recv()
to provide both the data and size of the data to read a packet into.

On success, update struct ofpbuf size inside netdev_class rx_recv
implementation and return 0. This moves logic from the caller.
On error a positive error code is returned, whereas previously
a negative error code was returned. This is a more common convention.

This patch should not have any behavioural changes.

This patch is in preparation for the netdev-linux variant of rx_recv()
making use of headroom in the struct ofpbuf * parameter to push a VLAN tag
obtained from auxdata.

Co-authored-by: Ben Pfaff <blp@nicira.com>
Signed-off-by: Simon Horman <horms@verge.net.au>

---
v5
* Updated documentation of rx_recv to text suggested by Ben Pfaff.

v4
* On error return a positive error code rather than a negative one
* Correct error in BSD changes: arg.retval rather than retval
  is the length of the data read in netdev_rx_bsd_recv_pcap().

v3
* On success, update struct ofpbuf size inside netdev_class rx_recv
  implementation and return 0. This moves logic from the caller.
* Document that it is recommended that the struct ofpbuf should provide
  extra tailroom for a VLAN header. This will be used by a subsequent patch.
* Update BSD code

v2
* First post