ofproto-dpif-xlate: Do not add more MPLS LSEs than the datapath can handle
This is an proposed enhancement to
"Implement OpenFlow support for MPLS, for up to 3 labels."

This makes the conservative assumption that the datapath should
not be asked to apply MPLS push actions such that the resulting
packet will have an MPLS label stack depth greater than the
datapath can accept in a match.

In future it may be possible to use recirculation to handle
such cases.

Signed-off-by: Simon Horman <horms@verge.net.au>