ofproto-dpif: Do not add more MPLS labels to mask than the datapath can handle
The key is sourced from the datapath so should not
have more labels than it can handle.

dpif-netdev supports as many LSEs as can fit in struct flow,
so it is safe to pass SIZE_MAX as the limit there.

This is an proposed enhancement to
"Implement OpenFlow support for MPLS, for up to 3 labels."

Signed-off-by: Simon Horman <horms@verge.net.au>